Moving ScalableOffset to scale.py
